+++ R/utils.R
@@ -4,7 +4,3 @@ utils::globalVariables(c(".", ":=", "%>%"))
 #' @export
 dplyr::`%>%`
 
-
-
-
-