+++ NEWS.md
@@ -7,6 +7,7 @@
 - Tarpuy: select sheet for field-book sketch
 - Tarpuy: create field-book only with the factor list
 - Tarpuy: column with [] are omitted in the field-book generation
+- CRAN comments: if (class(model) == "lmerMod") => if ( is(model, "lmerMod") 
 
 # inti 0.5.4